+++ packages/mongodb-memory-server-core/src/MongoMemoryServer.ts
@@ -9,6 +9,7 @@ import {
   authDefault,
   statPath,
   ManagerAdvanced,
+  Cleanup,
 } from './util/utils';
 import { MongoInstance, MongodOpts, MongoMemoryInstanceOpts } from './util/MongoInstance';
 import { MongoBinaryOpts } from './util/MongoBinary';
@@ -276,10 +277,17 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
 
     this.stateChange(MongoMemoryServerStates.starting);
 
-    await this._startUpInstance(forceSamePort).catch((err) => {
+    await this._startUpInstance(forceSamePort).catch(async (err) => {
       if (!debug.enabled('MongoMS:MongoMemoryServer')) {
-        console.warn('Starting the instance failed, enable debug for more information');
+        console.warn(
+          'Starting the MongoMemoryServer Instance failed, enable debug log for more information. Error:\n',
+          err
+        );
       }
+
+      this.debug('_startUpInstance threw a Error: ', err);
+
+      await this.stop(false); // still try to close the instance that was spawned, without cleanup for investigation
 
       this.stateChange(MongoMemoryServerStates.stopped);
 
@@ -326,8 +334,10 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   /**
    * Construct Instance Starting Options
    */
-  protected async getStartOptions(): Promise<MongoMemoryServerGetStartOptions> {
-    this.debug('getStartOptions');
+  protected async getStartOptions(
+    forceSamePort: boolean = false
+  ): Promise<MongoMemoryServerGetStartOptions> {
+    this.debug(`getStartOptions: forceSamePort: ${forceSamePort}`);
     /** Shortcut to this.opts.instance */
     const instOpts = this.opts.instance ?? {};
     /**
@@ -335,8 +345,16 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
      */
     let isNew: boolean = true;
 
+    // use pre-defined port if available, otherwise generate a new port
+    let port = typeof instOpts.port === 'number' ? instOpts.port : undefined;
+
+    // if "forceSamePort" is not true, and get a available port
+    if (!forceSamePort || isNullOrUndefined(port)) {
+      port = await this.getNewPort(port);
+    }
+
     const data: StartupInstanceData = {
-      port: await this.getNewPort(instOpts.port),
+      port: port,
       dbName: generateDbName(instOpts.dbName),
       ip: instOpts.ip ?? '127.0.0.1',
       storageEngine: instOpts.storageEngine ?? 'ephemeralForTest',
@@ -362,7 +380,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
         );
         const files = await fspromises.readdir(data.dbPath);
 
-        isNew = files.length > 0; // if there already files in the directory, assume that the database is not new
+        isNew = files.length === 0; // if there are no files in the directory, assume that the database is new
       }
     } else {
       isNew = false;
@@ -412,7 +430,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
       return;
     }
 
-    const { mongodOptions, createAuth, data } = await this.getStartOptions();
+    const { mongodOptions, createAuth, data } = await this.getStartOptions(forceSamePort);
     this.debug(`_startUpInstance: Creating new MongoDB instance with options:`, mongodOptions);
 
     const instance = await MongoInstance.create(mongodOptions);
@@ -455,9 +473,31 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   /**
    * Stop the current In-Memory Instance
    * @param runCleanup run "this.cleanup"? (remove dbPath & reset "instanceInfo")
+   *
+   * @deprecated replace argument with `Cleanup` interface object
    */
-  async stop(runCleanup: boolean = true): Promise<boolean> {
+  async stop(runCleanup: boolean): Promise<boolean>; // TODO: for next major release (9.0), this should be removed
+  /**
+   * Stop the current In-Memory Instance
+   * @param cleanupOptions Set how to run ".cleanup", by default only `{ doCleanup: true }` is used
+   */
+  async stop(cleanupOptions?: Cleanup): Promise<boolean>;
+  async stop(cleanupOptions?: boolean | Cleanup): Promise<boolean> {
     this.debug('stop: Called .stop() method');
+
+    /** Default to cleanup temporary, but not custom dbpaths */
+    let cleanup: Cleanup = { doCleanup: true, force: false };
+
+    // handle the old way of setting wheter to cleanup or not
+    // TODO: for next major release (9.0), this should be removed
+    if (typeof cleanupOptions === 'boolean') {
+      cleanup.doCleanup = cleanupOptions;
+    }
+
+    // handle the new way of setting what and how to cleanup
+    if (typeof cleanupOptions === 'object') {
+      cleanup = cleanupOptions;
+    }
 
     // just return "true" if there was never an instance
     if (isNullOrUndefined(this._instanceInfo)) {
@@ -467,26 +507,18 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
     }
 
     if (this._state === MongoMemoryServerStates.stopped) {
-      this.debug(`stop: state is "stopped", so already stopped`);
-
-      return false;
+      this.debug('stop: state is "stopped", trying to stop / kill anyway');
     }
 
-    // assert here, otherwise typescript is not happy
-    assertion(
-      !isNullOrUndefined(this._instanceInfo.instance),
-      new Error('"instanceInfo.instance" is undefined!')
-    );
-
     this.debug(
-      `stop: Stopping MongoDB server on port ${this._instanceInfo.port} with pid ${this._instanceInfo.instance.mongodProcess?.pid}` // "undefined" would say more than ""
+      `stop: Stopping MongoDB server on port ${this._instanceInfo.port} with pid ${this._instanceInfo.instance?.mongodProcess?.pid}` // "undefined" would say more than ""
     );
     await this._instanceInfo.instance.stop();
 
     this.stateChange(MongoMemoryServerStates.stopped);
 
-    if (runCleanup) {
-      await this.cleanup(false);
+    if (cleanup.doCleanup) {
+      await this.cleanup(cleanup);
     }
 
     return true;
@@ -498,9 +530,43 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
    * @throws If "state" is not "stopped"
    * @throws If "instanceInfo" is not defined
    * @throws If an fs error occured
+   *
+   * @deprecated replace argument with `Cleanup` interface object
    */
-  async cleanup(force: boolean = false): Promise<void> {
+  async cleanup(force: boolean): Promise<void>; // TODO: for next major release (9.0), this should be removed
+  /**
+   * Remove the defined dbPath
+   * @param options Set how to run a cleanup, by default `{ doCleanup: true }` is used
+   * @throws If "state" is not "stopped"
+   * @throws If "instanceInfo" is not defined
+   * @throws If an fs error occured
+   */
+  async cleanup(options?: Cleanup): Promise<void>;
+  async cleanup(options?: boolean | Cleanup): Promise<void> {
     assertionIsMMSState(MongoMemoryServerStates.stopped, this.state);
+
+    /** Default to doing cleanup, but not forcing it */
+    let cleanup: Cleanup = { doCleanup: true, force: false };
+
+    // handle the old way of setting wheter to cleanup or not
+    // TODO: for next major release (9.0), this should be removed
+    if (typeof options === 'boolean') {
+      cleanup.force = options;
+    }
+
+    // handle the new way of setting what and how to cleanup
+    if (typeof options === 'object') {
+      cleanup = options;
+    }
+
+    this.debug(`cleanup:`, cleanup);
+
+    // dont do cleanup, if "doCleanup" is false
+    if (!cleanup.doCleanup) {
+      this.debug('cleanup: "doCleanup" is set to false');
+
+      return;
+    }
 
     if (isNullOrUndefined(this._instanceInfo)) {
       this.debug('cleanup: "instanceInfo" is undefined');
@@ -513,8 +579,6 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
       new Error('Cannot cleanup because "instance.mongodProcess" is still defined')
     );
 
-    this.debug(`cleanup: force ${force}`);
-
     const tmpDir = this._instanceInfo.tmpDir;
 
     if (!isNullOrUndefined(tmpDir)) {
@@ -522,7 +586,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
       tmpDir.removeCallback();
     }
 
-    if (force) {
+    if (cleanup.force) {
       const dbPath: string = this._instanceInfo.dbPath;
       const res = await statPath(dbPath);
 
@@ -625,10 +689,13 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   /**
    * Generate the Connection string used by mongodb
    * @param otherDb add an database into the uri (in mongodb its the auth database, in mongoose its the default database for models)
+   * @param otherIp change the ip in the generated uri, default will otherwise always be "127.0.0.1"
+   * @throws if state is not "running" (or "starting")
+   * @throws if an server doesnt have "instanceInfo.port" defined
    * @returns an valid mongo URI, by the definition of https://docs.mongodb.com/manual/reference/connection-string/
    */
-  getUri(otherDb?: string): string {
-    this.debug('getUri:', this.state);
+  getUri(otherDb?: string, otherIp?: string): string {
+    this.debug('getUri:', this.state, otherDb, otherIp);
 
     switch (this.state) {
       case MongoMemoryServerStates.running:
@@ -644,7 +711,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
 
     assertionInstanceInfo(this._instanceInfo);
 
-    return uriTemplate(this._instanceInfo.ip, this._instanceInfo.port, generateDbName(otherDb));
+    return uriTemplate(otherIp || '127.0.0.1', this._instanceInfo.port, generateDbName(otherDb));
   }
 
   /**
@@ -670,7 +737,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
     this.debug(`createAuth: Creating Root user, name: "${this.auth.customRootName}"`);
     await db.command({
       createUser: this.auth.customRootName,
-      pwd: 'rootuser',
+      pwd: this.auth.customRootPwd,
       mechanisms: ['SCRAM-SHA-256'],
       customData: {
         createdBy: 'mongodb-memory-server',